EC Refactor with reed-solomon-simd 3.0.1  (#361)
The reed-solomon-simd 3.0.1 package is easier to FFI into because it only has a 2-byte alignment requirement.
Thus instead of FFIing into Parity's package, it is possible to directly FFI into this. It was also easier to just have the decode and encode erasure coding functions in the "bls" package.

It is not worth the busy work of separating the EC out from bls. Its probably the case that its even smarter to fold bandersnatch into a single "crypto" library for all FFIs so there is less thinking about N packages. Lets not do that either, we are not really actively working on bls OR bandersnatch internals anymore NOR should we engage in erasure coding interfaces.

The key interfaces of Decode and Encode are seen in the erasurecoding_test.go file. Encode generates all V shards but Decode requires exactly C of them AND their indices.

Previously, Stanley had a bit more complexity with 3d arrays and a lot of abstractions that were used exactly once -- I simplified with this new simpler interface. I believe we should use the diagrams in the Availability Specifier doc to reason about this instead of code.

There are only 4 CEs that connect to erasure coding

CE 137: Shard distribution
CE 138: Audit shard request
CE 139: Segment shard request
CE 140: same as above but with Justification request
Since we are not doing auditing this moment, and CE 140 is not being used yet, we need to focus on getting only ONE things correct next: CE 137. As far as I can tell CE 139 is working but we need to test more import and more exported segments with dev.

The Justification checking for CE137 has been disabled as of this commit. So we are now ready for rework on CDT Justification.
---------

Co-authored-by: Shawn Chung <122467037+yoyo2325@users.noreply.github.com>