PVM Logging cleanup; OnChain => Accessed; sub with slot/headerhash; LookupObject+PreimageObject support for Accessed (#454)
Addresses #451 

## PVM and General Authoring vs Validating
The PVM Package should not make logging decisions.  The caller should set the PVMContext, and this is used to set the module of "vm.logging" (which is the module).  Then EnableModule, also outside pvm, decides whether something appears in the logs for Debug+Trace along with logging level.

The SetLogging is renamed SetPVMContext to make it clear for each of { Auth/Refine/Auditing, AccumulateTransfer }.
There were many `if vm.logging != ""` that was hard to reason about.   Those are gone now.

Similarly, there were many `if sd.Authoring { ... }` that are now just `log.Debug/Trace(sd.Authoring, ...)` 

This way, for all 4 situations, you can EnableModule.